Move pipeline admin endpoints to router
